Added Influence vel. ground decay and Debug Flying
influence velocity now decays quickly when on the ground and slightly slower when in the air. allows for the player to conserve a little speed if they jump after landing from a launcher.
debug flying allows for the player to fly quickly, making testing later parts of a level easier.